more specific requires in rails helper
