--- spec/support/api_helper.rb
@@ -1,11 +1,0 @@
-module ApiHelper
-  include Rack::Test::Methods
-
-  def app
-    Rails.application
-  end
-end
-
-RSpec.configure do |config|
-  config.include ApiHelper, type: :api # apply to all spec for apis folder
-end
