+++ spec/support/config/capybara.rb
@@ -2,6 +2,7 @@ require "capybara/rails"
 require "capybara/rspec"
 require "capybara-screenshot/rspec"
 require "selenium/webdriver"
+require_relative "../download_helpers"
 
 # not used unless you swap it out for selenium_chrome_headless_in_container to watch tests running in docker
 Capybara.register_driver :selenium_chrome_in_container do |app|
