+++ spec/views/templates/email_templates_spec.rb
@@ -1,6 +1,8 @@
 require "rails_helper"
 
 RSpec.shared_examples "compares opening and closing tags" do
+  include TemplateHelper
+
   it "validates html tags" do
     file_content = File.read(Rails.root.join(file_path))
     tags_are_equal = validate_closing_tags_exist(file_content)
